+++ src/test/java/io/personium/test/io/model/imple/fs/DavMetadataFileTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.personium.core.model.impl.fs;
+package io.personium.test.io.model.imple.fs;
 
 import static java.lang.ClassLoader.getSystemResourceAsStream;
 import static org.hamcrest.CoreMatchers.instanceOf;
@@ -48,6 +48,7 @@ import org.powermock.modules.junit4.PowerMockRunner;
 
 import io.personium.core.PersoniumCoreException;
 import io.personium.core.PersoniumUnitConfig;
+import io.personium.core.model.impl.fs.DavMetadataFile;
 import io.personium.test.categories.Unit;
 
 /**
